+++ main/src/addins/MonoDevelop.AspNetCore/MonoDevelop.AspNetCore.Scaffolding/ScaffolderWizardPageBase.cs
@@ -1,0 +1,65 @@
+﻿//
+// Scaffolder.cs
+//
+// Author:
+//       jasonimison <jaimison@microsoft.com>
+//
+// Copyright (c) 2019 Microsoft
+//
+// Permission is hereby granted, free of charge, to any person obtaining a copy
+// of this software and associated documentation files (the "Software"), to deal
+// in the Software without restriction, including without limitation the rights
+// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
+// copies of the Software, and to permit persons to whom the Software is
+// furnished to do so, subject to the following conditions:
+//
+// The above copyright notice and this permission notice shall be included in
+// all copies or substantial portions of the Software.
+//
+// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
+// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
+// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
+// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
+// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
+// THE SOFTWARE.
+using MonoDevelop.Components;
+using MonoDevelop.Ide.Gui.Wizard;
+using Xwt;
+
+namespace MonoDevelop.AspNetCore.Scaffolding
+{
+	abstract class ScaffolderWizardPageBase : WizardDialogPageBase
+	{
+		string subSubTitle;
+		protected ScaffolderWizardPageBase (ScaffolderArgs args)
+		{
+			Args = args;
+			CanGoBack = true;
+			CanGoNext = true;
+		}
+
+		public string SubSubTitle {
+			get => subSubTitle;
+			protected set => subSubTitle = value;
+		}
+
+		public ScaffolderArgs Args { get; }
+
+		protected override Control CreateControl ()
+		{
+			var icon = new Xwt.ImageView (StockIcons.Information);
+			var mainBox = new VBox { Spacing = 0 };
+			var label = new Label (subSubTitle);
+			label.Font = label.Font.WithSize (18);
+			mainBox.PackStart (label, margin: 30);
+			var separator = new HSeparator ();
+			mainBox.PackStart (separator);
+
+			mainBox.PackStart (GetMainControl (), margin: 20);
+			return new XwtControl (mainBox);
+		}
+
+		protected abstract Widget GetMainControl ();
+	}
+}
